requirements.txt
Django==1.9.2
decorator==4.0.9
dj-cmd==0.5
dj-database-url==0.4.0
django-bootstrap3==6.2.2
django-dual-authentication==1.0.0
django-easy-select2==1.3.2
django-enumfield==1.3b2
django-js-reverse==0.6.1
django-mptt==0.8.0
django-mptt-admin==0.3.3
django-registration==2.0.4
gunicorn==19.4.5
ipython==4.1.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
